+++ src/pages/Dashboard/Dashboard.jsx
@@ -1,9 +1,9 @@
-import Sidebar from '../components/Sidebar'
 
 const Dashboard = () => {
   return (
     <>
-      <Sidebar />
+      <h1>Dashboard</h1>
+      <p>Bienvenido al dashboard de la aplicación</p>
     </>
   )
 }