+++ src/pages/Supply/Supply.jsx
@@ -1,0 +1,11 @@
+
+const Supply = () => {
+  return (
+    <>
+      <h1>Supply</h1>
+      <p>Bienvenido al Supply de la aplicación</p>
+    </>
+  )
+}
+
+export default Supply