+++ CHANGELOG.md
@@ -1,12 +1,13 @@
-## <small>0.0.7 (2023-05-16)</small>
+## <small>0.0.8 (2023-05-16)</small>
 
+* build: :construction: add dashboard ([32edbb3](https://github.com/Daintz/SenaOnPrintingFrontend/commit/32edbb3))
+* build: :hammer: separate config react router dom in other folder ([4045acb](https://github.com/Daintz/SenaOnPrintingFrontend/commit/4045acb))
 * feat: :sparkles: add method post and get ([87a8369](https://github.com/Daintz/SenaOnPrintingFrontend/commit/87a8369))
 * feat: :sparkles: add react router dom ([7b5bfa0](https://github.com/Daintz/SenaOnPrintingFrontend/commit/7b5bfa0))
 * chore: :construction: in-process api consumption ([bfabba9](https://github.com/Daintz/SenaOnPrintingFrontend/commit/bfabba9))
 * chore: :construction: update page information ([f189063](https://github.com/Daintz/SenaOnPrintingFrontend/commit/f189063))
 * chore: :tada: project structure ([328a6b4](https://github.com/Daintz/SenaOnPrintingFrontend/commit/328a6b4))
 * chore: :wrench: add hook ([d5f865e](https://github.com/Daintz/SenaOnPrintingFrontend/commit/d5f865e))
-* build: :construction: add dashboard ([32edbb3](https://github.com/Daintz/SenaOnPrintingFrontend/commit/32edbb3))
 * test: :test_tube: test husky hook ([3bf1c70](https://github.com/Daintz/SenaOnPrintingFrontend/commit/3bf1c70))
 * test: :white_check_mark: deploy in vercel ([6fe892f](https://github.com/Daintz/SenaOnPrintingFrontend/commit/6fe892f))